Version 0.2.9
Zahlungsverfolgung:

—Suche erweitert:
a. Wenige Ziffern der Beträge/RG-Nr/Gewichte reichen aus um
Datenbankeinträge zu finden
b. Nun auch nach Zeitspanne filterbar (z.B. offene RG zwischen Tag X
und Tag Y)
c. Es ist jetzt egal in welchen Kalender man ein Datum einträgt

—Excel-Exportfunktion für Tabellen eingebaut

—Gesamtgewicht angezeigter Rechnungen nun sichtbar

—Anzahl angezeigter Rechnungen nun sichtbar

Generell:
—Operationen_JTextField-Klasse unterstützt jetzt Summieren von
Datanbankspalten und normalen Tabellenspalten

+++ ARSLANKEBAPVerwaltungssoftware/src/myPackage/KundenPK.java
@@ -50,7 +50,7 @@ public class KundenPK implements Serializable {
     @Override
     public int hashCode() {
         int hash = 0;
-        hash += (int) kundennummer;
+        hash += kundennummer;
         hash += (kundenname != null ? kundenname.hashCode() : 0);
         return hash;
     }
